requirements.txt
Django==4.2.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.0
python-dotenv==1.0.0
Pillow==10.2.0
